Refactored the code,added swagger config and removed some boilerlate code

+++ src/main/java/com/myipl/domain/entity/IPLMatchWinner.java
@@ -11,13 +11,10 @@ import javax.persistence.Id;
 @Entity
 public class IPLMatchWinner implements Serializable {
 
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	private String match1Winner;
 
@@ -25,11 +22,11 @@ public class IPLMatchWinner implements Serializable {
 
 	private LocalDate matchDate;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
